requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-cloud-storage
ultralytics
opencv-python
numpy
httpx
python-multipart
